SWARM-436: Using a server side solution to generate ZIP
Motivation
----------
Safari does not support generating a file using pure JavaScript.
A good solution would be to make this generation server-side so it can be consumed by third parties

Modifications
-------------
The JS libraries that were used to generate the ZIP file were removed and now it's a simple redirect to
http://swarmgen-forge.rhcloud.com/generator, where a JAX-RS Resource consumes the passed parameters and
returns a ZIP file

Result
------
This will redirect to the server-side REST endpoint instead of creating the ZIP file locally.